104 Support `delete_all` on QuerySet
THe usual `delete` method on QuerySet instantiates objects and destroys them one by one, ensuring callbacks and validations are run. But for performance reasons, we may want to directly delete objects at the database level, say with SQL queries. `delete_all` is a pass through method that passes the filter criteria as-is to the adapter and lets the adapter handle mass deletion.

This commit also cleans up the definition of `delete()` instance method on Entity:
* It returns the deleted object in a frozen state (with `is_destroyed` set to True)
* It raises ObjectNotFoundError if the object was not found in the database

Closes #104