121 Eliminate explicit definitions of Models
* Enrich meta options supported in `Meta` inner class of Entity
* Models will be synthesized by Provider in `get_model` method
* Supported options are `schema_name`, `provider`, and `order_by`
	* `schema_name` is the name of the table/index/document in the database
	* `provider` is the provider name as defined in configuration
	* `order_by` is a list of column names used to order the results automatically

This change means that an explicit Model cannot be provided as an option. All aspects that need to be controlled will be provided as options in Entity Meta class.
Deeper support for changing column names in db will be provided in the future, if required, by additional options in Entity Meta class.

Closes #121