120 Replace `paginate()` with `limit()` and `offset()` methods
It is better to clearly demarcate `limit` and `offset` values, than have one `paginate` method. Also, `Pagination` is a concept that applies usually to how records are displayed across pages, and does not necessarily have anything to do with results or databases. So this PR breaks the `paginate` method into `limit` and 'offset`, and also replaces `Pagination` class with `ResultSet` class.

Note that the parameters sent into Use Cases continue to be `page` and `per_page` which is consistent with requests over HTTP.

Closes #120 and #132