Remove Entity Hooks
Entity hooks like `pre_save()` and `post_save()` tie up an Entity class
with underlying persistence. Any code that should be executed before saving
an entity should be done in Entity's Constructor, and code in `post_save()`
should be handled as part of a combination of events and Application Service code.