Introduce support for Data Transfer Objects (#192)
We can now introduce DTOs under an aggregate and transform Aggregate/Entity data to be transferable over the wire. These DTOs are discoverable as part of the domain because they are registered with the domain.

It is possible to simply declare them as classes, and not register them with the domain at all. But then the DTO class will have to shared and imported by all objects that use the DTO (ex. Application Services, Message Brokers, etc.) We can remove the functionality if that turns out to be the usual case.