Add support for persisting Events
This commit contains changes that persist the event into a log table
called `event_log`, before publishing it to brokers. This provides
a chance to recover from errors and allows consumers to use a periodic
`fetch since last used` query.

It is also good to generally have this table around for debugging
purposes. The make up of this table may change when EventSourcing support
comes to Protean.