Limit number of entity objects fetched on loading Aggregate
Introduce a config attribute `AGGREGATE_CHILDREN_LIMIT` to control
the number of entity records loaded from the database. By default,
the limit value is set to 100.

The attribute can be overridden in application configs, but
this means that more objects will be loaded into memory on loading
the aggregate. So this attribute value should be treated with caution.