Switch to `make test` as default
Bringing down the services and up again before running tests does not
work in our case because Elasticsearch takes a bit of time to start up
completely. It is better to keep all make commands clear of each other,
and only ensure docker services are up when running full tests.