Simplification of Container structure - Step 1
This commit is preparing Container to be the universal data container for
all Protean elements. As a first step, Views have been made compliant with
the new Container structure.

The change should be largely invisible to rest of the elements, but further
refactorings will produce large scale changes in the codebase.