Move declared_fields out of meta_ in Container
This is first of many commits to delink `meta_` object from holding
field references. This commit introduces the change to Container class,
with a special `fields()` module method that returns a dict of defined
fields.

ContainerMeta has been altered to populate the fields in the order
specified. Next steps will be to migrate entity class (and aggregate
class) to this structure.