Manage concurrency with aggregate versions
This commit introduces a `_version` attribute to event-sourcing aggregates.
New aggregates start with version 0, which is incremented with every
subsequent message in the same stream.

When a new message is added to the stream, the expected version is compared
to the version of the aggregate. When they don't match, the situation is
treated as a concurrency conflict and the write operation is failed with
`ExpectedVersionError`.