Add optimistic concurrency to Aggregate updates
This commit introduces a `_version` field to BaseAggregate for
managing optimistic concurrency. This field is added automatically
to all aggregates, but does not reflect in ths aggregate's list of
declared fields. The version number starts by default as -1, with
the first aggregate version being 0. It is incremented and persisted
automatically with every change to the aggregate.

This commit also introduces `declared_fields` as a util method for
reflection.