Process events inline in Sync Mode
This commit introduces the ability to process events synchronously
when protean is running in Sync mode.

As part of this change, events need to be associated with a stream
or aggregate class to be valid. They can be associated explicitly,
or they are autmatically associated when consumed in event handlers
(with the `handle` method) or event sourced aggregates (with the
`apply` method).