Process commands inline in Sync Mode
This commit introduces the ability to process commands synchronously
when protean is running in Sync mode.

As part of this change, commands need to be associated with a stream
or aggregate class to be valid. They can be associated explicitly,
or they are automatically associated when consumed in command handlers
(with the `handle` method).