Handle repetitive updates on ES Aggregates in UoW
 UoW preserves an Identity Map of all objects that are encoutered
during its lifecycle, in `get` or `add` at repository level, or `save`,
`create` and `update` at dao level. The items in this map are checked
for new events on UoW commit.

If a process retrived the same aggregate more than once, the item
in the map object was being overwritten. With this commit, an object
that is already loaded into the identity map is retrived on `get` so
that events (a.k.a all changes as part of the process) are chained
together.