Add support for schema switching
This commit contains changes to switch schemas before establishing
a session to the database. Earlier, Protean was supplying the metadata
containing the schema only once, when the engine was initialized. This
commit specifies the schema just before establishing the session.

A thorough revamp is required to streamline the construction of
models and their usage of sessions, to be able to improve multi-
tenancy support in Protean.