Enhance `domain.init` as the way to activate a domain
This means that we can do with dynamic initializations that happen
on first invocation in providers, brokers, caches, and event stores.

A new method `reinitialize` has been introduced into the domain object
to be used when explicit reinitialization is necessary.

This commit also includes a new test category `WITH_COVERAGE`.
Github Actions has been changed to use this mechanism to run tests.