Resolve references when initializing Domain (#419)
This commit moves resolving references, as well as validating domain constructs
to domain initialization, not activation. This ensures they happen one time, at
the beginning of domain setup. If they are in activation, the resolution would
be triggered every time a domain context is activated, which would be every time
the domain processes a request!

This commit also has the following fixes:
* Ensure Value Object field's class resolution works
* All Events and Commands need to be associated with an Aggregate or a stream
* Validation to ensure events are not associated with multiple event sourced aggregates
* Allow Abstract events and commands to be defined without an Aggregate or a stream
* Scope `setup_db` fixtures to Module level (Having them at `session` level
was resulting in the wrong context being popped)
* Move `fetch_element_cls_from_registry` under Domain object to avoid using `current_domain`
* Pass `domain` during class resolution to avoid using `current_domain`