Remove `via` param for associations
This may be reintroduced in the future, but with better implementation.
It is unclear if we will need to custom connect an aggregate and its
entity at this stage, because of the way aggregates enclose entities
completely.

The only reason why this customization will be necessary
is when adapting to existing data structures. More enhancements will
be needed to enable full support.