Run all invariants on init and attribute changes
When an aggregate or entity object is initialized, or its attribute value
is changed, this commit ensures that all `clean()` methods (which in turn
run all invariant validations) are run across the aggregate.

This commit also adds `atomic_change` context manager for scenarios
where multiple aggregates in an aggregate need to be changed before
it becomes valid again.