Allow control of automatic identifier field addition
Entities and Aggregates in Protean are automatically augmented with
an identity field (named `id`), when one has not been explicitly provided.
Until now, the assumption was the abstract aggregates will not need to
have identifier fields, which was incorrect.

This commit allows control of automatic addition of identity field with
an explicit flag named `auto_add_id_field`. With this change, even
abstract aggregates can have identity fields.