Track aggregate cluster to which elements belong
This commit introduces a meta attribute named `aggregate_cluster` in entities,
aggregates, event sourced aggregate, events and commands, to track the
cluster they belong to. The aggregate is necessary to derive a stream name
if an explicit stream is not defined at the level of an event or command.

The aggregate cluster assignment is performed during `domain.init()`,
so all test files are updated to invoke `init()` after element registration.