Remove `provider` meta option from Entities
Entities are to be persisted in the same persistence store as their aggregates.
This way, the aggregate cluster is kept together and transaction boundaries
can be honored. This commit removes the explicit `provider` option from Entities.

A new method has been introduced at the domain level, that is invoked by the
`domain.init()` method, that takes care of setting `provider` for each entity.