Only initialize adapters when `domain.init()` is called
`Domain.__init__()` was calling `_initialize()` which was again invoked in `domain.init()`.
The assumption was that in default adapters are initialized in `__init__`, which is incorrect.
Adapters should be initialized only on `domain.init()`, by when all configurations would
be complete. Invoking `_initialize` twice is unnecessary and can have unintended consequences.

Also contains documentation on views.