Eventing related changes:
- Introduce `Metadata` Valueobject within events to store event metadata,
instead of relying on metadata in Messages
- Introduce `payload` property in Event to retrieve event payload / aggregate
data. This is necessary because there wiil be three major parts to each event
eventually - data (payload), metadata of event, and headers
- Introduce a `from_events` factory method to encapsulate the aspect of
initializing an aggregate from the first event, applying the event on it,
and subsequently applying the rest of the events. This keeps the logic of
aggregate initialization in a single place.
- Pass `Event` object to `_apply` method in Event Sourced Aggregates, instead
of a `Dict`. We were passing a Message payload, contents of which is not necessary
for aggregate data attributes.
- Clean up snapshot fetching and aggregate initialization because of the above
changes