Remove domain-level event `raise_` functionality
Events, at this point, should always be connected to aggregates,
and raised from within them as part of business logic. Directly
adding events at the domain level was helper functionality, but
with the addition of metadata and headers at the event level,
events have a lot more contextuality. They cannot be raised or
appended to event store directly.

We will need to revisit this functionality if we need Domain
Events (or some variation thereof) that are not connected to
aggregates.