Introduce `id` into Event metadata
`id` serves as the primary identifier of an event. It's format is
<domain-name>.<event-class-name>.<event-version>.<aggregate-id>.<aggregate-version>

This is applicable for Event Sourced Aggregates, and may undergo slight modifications
for regular Aggregates.

The changes in tests are largely because events were being added to the event store
directly. We cannot add plain events anymore because their metadata is being
enriched in the Aggregate's `raise_` method.