Manage handlers by event/command type
We were basing the logic on retrieving handlers on the fully
qualified name of the event or command class. This commit moves
the logic to be on the `__type__` attribute of event/command
classes.

This is one of the steps necessary to be able to handle events/
messages from other systems, where we may not have an event
registered in the domain. It is also necessary to be able to
support multiple versions of event and command classes.