Nest Event Sourcing functionality within Base Aggregate
So far, Event Sourced Aggregate was a segarate domain element with its own class
and factory. This causes a few issues:
- A separate API for event sourced aggregates
- Duplication of code between aggregate types
- When a bug is fixed in one place, the other has to be take care too
- Dissimilarities in functionality (Event sourced aggregates did not support invariants
  and entity nesting)

This commit merges the functionality and introduces an option called `is_event_sourced`
on BaseAggregate to control event sourcing behavior. The repository will remain as-is
for now, because that's where event sourced aggregates primarily differ from regular
aggregates. This will need to be revisited if we want to support custom event sourced
repositories.