Simplify Broker and Subscriber functionality
Previously, brokers in Protean were the primary conduits for events. So a lot of
their functionality included automatic translation of events to messages and back,
and handling Protean elements.

This commit simplifies broker functionality, and retains one primary responsibility:
communicating with a broker (publish and receive).

`Broker.publish` is used to push messages to a broker.
Subcribers connected to channels receive messages from broker.